backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2 